Fix small brain mistake

+++ ft_strncmp.c
@@ -14,10 +14,10 @@
 
 int	ft_strncmp(const char *s1, const char *s2, size_t n)
 {
-	int i;
+	size_t i;
 
 	i = 0;
-	while ((s1[i] || s2[i]) && i < (int)n)
+	while ((s1[i] || s2[i]) && i < n)
 	{
 		if ((unsigned char)s1[i] != (unsigned char)s2[i])
 			return ((unsigned char)s1[i] - (unsigned char)s2[i]);
